- add some benchmarks  - port to osx  - a few other cleanups
git-svn-id: svn+ssh://smote/code/powerdb/readingdb@5217 16e81406-0b52-4fc8-9ac4-c3f00167659a

+++ c6/stats-sock.c
@@ -12,9 +12,9 @@
 #include "logging.h"
 #include "stats.h"
 
-static short sock = -1;
+static int sock = -1;
 
-void stats_init(short p) {
+void stats_init(unsigned short p) {
   struct sockaddr_in6 dest = {
     .sin6_family = AF_INET6,
     .sin6_addr = IN6ADDR_LOOPBACK_INIT,
